pyproject.toml
[project]
name = "rpi-monitor"
version = "0.1.0"
description = "Monitoramento de métricas do Raspberry Pi 4 com persistência em PostgreSQL"
requires-python = ">=3.11"
dependencies = [
    "psutil>=5.9",
    "psycopg[binary]>=3.1",
    "python-dotenv>=1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["rpi_monitor*"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"
